requirements.txt
setuptools
wheel
pyserial
influxdb
requests